Fix for issue #76 - on simple SQLite JDBC dialects, uses SQLite specific technique to fetch the auto-generated primary key after an insert.
Summary of changes:

A SqlDialect is always provided for the database - the default is
a generic dialect.
Much of the functionality in insertAndGenerateKey is refactored into the
SqlDialect, via two functions - prepareStatement() and generatedKey().
The default implementation of these matches the original code in Dml.kt.
A new class SimpleSqliteDialect is implemented in ktorm-support-sqlite.
This overrides those functions with code that works on SQLite.

The effect is that if the database dialect is set to an instance of
SimpleSqlDialect then the package works on Android for the use case
of insertAndGenerateKey().